Continue On Blue-Virtual-Dragon 4.9.232-BvD-Fenix_r15+
Merged CAF TAGS = "LA.UM.8.6.2.r1-06900-89xx.0" "LA.UM.8.6.2.r1-07100-89xx.0"
"LA.UM.8.6.2.r1-07600-89xx.0" "LA.UM.8.6.2.r1-07800-89xx.0" "LA.UM.8.6.2.r1-08100-89xx.0" On kernel.lnx.4.9.r27-rel

* New MODS Fixed And Working =
- SchedTune Assit. X
- CPUSET ASSIST. X
- Devfreq_Boost. X
- SchedTune Boost. X
- Cpu_Input_Boost. X

- Defconfig Updated
- Close To 4.9.233 Synced With Last 4.9-queue linux-stable-queue.

Clean UP Kernel Sources.

+++ drivers/mtd/spi-nor/spi-nor.c
@@ -733,7 +733,7 @@ static int spi_nor_unlock(struct mtd_info *mtd, loff_t ofs, uint64_t len)
 static int spi_nor_is_locked(struct mtd_info *mtd, loff_t ofs, uint64_t len)
 {
 	struct spi_nor *nor = mtd_to_spi_nor(mtd);
-	int ret;
+	ssize_t ret;
 
 	ret = spi_nor_lock_and_prep(nor, SPI_NOR_OPS_UNLOCK);
 	if (ret)
